% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_data.R
\name{generate_summary_data}
\alias{generate_summary_data}
\title{Generate Simulated Data for Mendelian Randomization Analysis}
\usage{
generate_summary_data(
  theta,
  m,
  Rbb,
  Ruv,
  Rnn,
  Nxy,
  Hxy,
  LD = "identify",
  non.zero.frac,
  pleiotropy.frac = 0,
  pleiotropy.var = 0.5,
  effect.dis = "normal",
  pleiotropy.dis = "uniform"
)
}
\arguments{
\item{theta}{An (px1) vector of causal effects.}

\item{m}{The number of instrumental variables (IVs).}

\item{Rbb}{An (pxp) correlation matrix of genetic effects.}

\item{Ruv}{An ((p+1)x(p+1)) correlation matrix of residuals in outcome and exposures; the outcome is the first one.}

\item{Rnn}{An ((p+1)x(p+1)) correlation matrix of sample overlap; the outcome is the first one.}

\item{Nxy}{An ((p+1)x1) vector of GWAS sample sizes; the outcome is the first one.}

\item{Hxy}{An ((p+1)x1) vector of heritabilities; the outcome is the first one.}

\item{LD}{An (mxm) correlation matrix of the IVs or "identify" indicating independent IVs.}

\item{pleiotropy.frac}{A number in [0,0.5) indicating the fraction of IVs affected by pleiotropy.}

\item{pleiotropy.var}{A number in [0,+∞) indicating the variance attributed to pleiotropy.}

\item{effect.dis}{Distribution of genetic effects: "normal" (default), "uniform", or "t" distribution (with degree of freedom 5).}

\item{pleiotropy.dis}{Distribution of pleiotropy effects: "normal" (default), "uniform", or "t" distribution (with degree of freedom 5).}

\item{zero.frac}{An (px1) vector with all entries in (0,1]; each entry is the probability of deltaj such that betaj=betaj'*deltaj.}
}
\value{
A list containing simulated GWAS effect sizes for exposures (bX), their standard errors (bXse),
        the GWAS effect size for the outcome (by), its standard error (byse), and the pleiotropy effects (pleiotropy).
}
\description{
This function generates simulated data for Mendelian Randomization (MR) analysis,
considering genetic effects, estimation errors, and horizontal pleiotropy.
It allows for different distributions of genetic effects and pleiotropy,
and accommodates both independent and correlated instrumental variables (IVs).
}
\note{
This function requires the `MASS`, `mvtnorm`, `stats`, and `Matrix` packages.
}
