% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBCEE.R
\name{MRBCEE}
\alias{MRBCEE}
\title{Mendelian Randomization using Bias-correction Correlated Estimating Equation.}
\usage{
MRBCEE(
  by,
  bX,
  byse,
  bXse,
  LD,
  Rxy,
  cluster.index,
  Nmin = F,
  tauvec = seq(3, 50, by = 2),
  max.iter = 100,
  max.eps = 0.001,
  ebic.gamma = 1,
  reliability.thres = 0.5,
  rho = 2,
  maxdiff = 1.5,
  parametric = F,
  sampling.time = 0,
  sampling.frac = 0.5,
  theta.ini = F,
  gamma.ini = F
)
}
\arguments{
\item{by}{A vector (n x 1) of GWAS effect sizes for the outcome.}

\item{bX}{A matrix (n x p) of GWAS effect sizes for p exposures.}

\item{byse}{A vector (n x 1) of standard errors for the GWAS effect sizes of the outcome.}

\item{bXse}{A matrix (n x p) of standard errors for the GWAS effect sizes of the exposures.}

\item{LD}{A matrix representing the linkage disequilibrium (LD) among instrumental variables.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix including p exposures and the outcome. Outcome should be the last column.}

\item{cluster.index}{A vector indicating the cluster membership for each instrumental variable. This is used in standard error estimation.}

\item{Nmin}{Optional; the minimum sample size for the GWAS if not provided, defaults to the number of instrumental variables.}

\item{tauvec}{A vector of tuning parameters for penalizing horizontal pleiotropy in the IPOD algorithm.}

\item{max.iter}{The maximum number of iterations allowed for convergence of the causal effect estimates.}

\item{max.eps}{The tolerance level for convergence; iteration stops when changes are below this threshold.}

\item{ebic.gamma}{The penalty factor for extended Bayesian Information Criterion (eBIC) adjustments on pleiotropy.}

\item{reliability.thres}{A threshold on bias-correction term, defaults to 0.5.}

\item{rho}{The penalty multiplier used in the ADMM algorithm within the IPOD framework.}

\item{maxdiff}{The maximum allowed difference ratio between iterative causal estimates and initial estimations for stabilization.}

\item{sampling.time}{The number of subsampling iterations used to estimate the standard error of the causal effect estimate. Defaults to 1. When set to 1, a sandwich formula is applied for the estimation.}

\item{sampling.frac}{The fraction of the data to be used in each subsampling iteration. Defaults to 0.5, meaning that 50% of the data is used in each iteration.}

\item{theta.ini}{Initial estimates for the causal effects; defaults to FALSE, indicating automatic initialization.}

\item{gamma.ini}{Initial estimates for horizontal pleiotropy effects; also defaults to FALSE for automatic setup.}
}
\value{
A list containing detailed results of the analysis, including estimated causal effects, pleiotropy effects, their respective standard errors, and Bayesian Information Criterion (BIC) scores, among other metrics.
}
\description{
Detailed description of the function goes here.
}
\note{
Requires the `Rcpp`, `RcppArmadillo`, and `varbvs` packages for computational efficiency and modeling.
}
